% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_algorithm.R
\name{.match_algorithm}
\alias{.match_algorithm}
\title{The matching algorithm}
\usage{
.match_algorithm(
  splist_class,
  max_distance,
  progress_bar = FALSE,
  keep_closest = TRUE,
  genus_fuzzy = TRUE,
  grammar_check = FALSE
)
}
\description{
The matching algorithm
}
\keyword{internal}
