% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook_distribution.R
\docType{data}
\name{redbook_distribution}
\alias{redbook_distribution}
\title{redbook_distribution Dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 10957 rows and 7 columns.
}
\usage{
redbook_distribution
}
\description{
This dataset, \code{redbook_distribution}, provides information on the
distribution of plant species listed in the Red Book of Endemic Plants
of Peru. It is a tibble (data frame) containing seven variables:
}
\details{
\describe{
\item{possible_names}{Character vector with possible names for each taxon.}
\item{taxon_name}{Character vector with the taxonomic names of the species.}
\item{accepted_name}{Character vector with the accepted names for each taxon.}
\item{accepted_name_author}{Character vector with the authors of the accepted names.}
\item{accepted_family}{Character vector with the accepted family for each taxon.}
\item{taxonomic_status}{Character vector indicating the taxonomic status (e.g., "Accepted", "Synonym") of each taxon.}
\item{level2_code}{Character vector with the level 2 code for the distribution of each taxon.}
}

The \code{redbook_distribution} dataset is a tibble (data frame) that contains the
distribution information for a subset of taxa listed in the Red Book of Endemic
Plants of Peru.
}
\examples{
\dontrun{
  head(redbook_distribution)
}

}
\seealso{
\href{https://revistasinvestigacion.unmsm.edu.pe/index.php/rpb/issue/view/153}{Red Book of Endemic Plants of Peru}
}
\keyword{datasets}
