% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbookperu_names.R
\docType{data}
\name{redbook_names}
\alias{redbook_names}
\title{Red Book of Endemic Plants of Peru Database}
\format{
A data frame with 10 columns and multiple rows.
\describe{
\item{possible_names}{Possible alternative names for the species.}
\item{accepted_name}{Accepted name of the species.}
\item{accepted_family}{Accepted family of the species.}
\item{accepted_name_author}{Author(s) of the accepted name.}
\item{taxonomic_status}{Taxonomic status of the species (e.g., Accepted, Synonym, etc.).}
\item{taxon_name}{Full taxonomic name of the species.}
}
}
\source{
The data has been extracted and compiled from the Red Book of Endemic Plants of Peru,
and subsequently standardized with the assistance of TNRS and WCVP.
}
\usage{
redbook_names
}
\description{
This database contains taxonomic information of endemic plant species from Peru,
including possible names, accepted names, accepted family, name authors, taxonomic status, and full taxon name.
}
\details{
This database serves as the primary reference for conducting partial and exact
searches of species listed in the Red Book of Endemic Plants of Peru. The information
in this database has been standardized using the Taxonomic Name Resolution Service (TNRS)
and the World Checklist of Vascular Plants (WCVP) to ensure consistency and accuracy.
}
\examples{
data(redbook_names)

}
\keyword{Peru,}
\keyword{TNRS,}
\keyword{WCVP}
\keyword{dataset,}
\keyword{endemic}
\keyword{plants,}
\keyword{taxonomy,}
