% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook_data.R
\docType{data}
\name{redbook_position}
\alias{redbook_position}
\title{List of the number positions of the first 3 letters of the species name in
the redbook_tab}
\format{
A data frame with 978 observations on the following 3 variables:
\describe{
\item{position}{A character vector. The position of the first three letters of the species name in the \code{redbook_tab}.}
\item{triphthong}{A character vector. The first three letters of the species name in the \code{redbook_tab}.}
\item{genus}{A character vector. The corresponding genus name.}
}
}
\usage{
redbook_position
}
\description{
The 'redbook_position' reports the
position (in term of number of rows) of the first three letters (triphthong)
for the plant names stored in the variable 'accepted_name' of the table
'redbook_tab'. This indexing system speeds up of the search on the
largest list using the package.
}
\details{
Positions of Species Names in The Red Book of Endemic Plants of Peru

The \code{redbook_position} dataset provides the positions of the first three letters of each species name listed in the \code{redbook_tab}.
}
\examples{

data("redbook_position")
head(redbook_position)

}
\keyword{datasets}
