% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redbook_data.R
\docType{data}
\name{redbook_sps_class}
\alias{redbook_sps_class}
\title{List of Species Names in redbook_sps_class Separated by Category}
\format{
A data.frame with the following columns:
\describe{
\item{species}{A character vector. The full species name.}
\item{genus}{A character vector. The genus of the species.}
\item{epithet}{A character vector. The specific epithet of the species.}
\item{input_subspecies_epitheton}{A character vector. The infraspecific epithet of the species, if applicable.}
\item{rank}{A character vector. The taxonomic rank (e.g., "species", "subspecies", "variety").}
\item{subspecies}{A character vector. The subspecies name, if applicable.}
\item{variety}{A character vector. The variety name, if applicable.}
\item{hybrid}{A character vector. Indicates if the species is a hybrid.}
\item{id}{A character vector. The ID of the species in the \code{redbook_tab}.}
}
}
\usage{
redbook_sps_class
}
\description{
The \code{redbook_sps_class} dataset includes all species names separated by genus,
epithet, author, subspecies, variety, and their position (ID) in the
\code{redbook_tab}.
}
\examples{

data("redbook_sps_class")
head(redbook_sps_class)

}
\keyword{datasets}
